CSRF token funcs updated
