+++ src/utils/auth.js
@@ -1,4 +1,18 @@
 import { targetURL, httpConfig, USER_AUTH_URL, URLS_USER_AUTH } from "utils/networkConfigs";
+import { setCookie } from "utils/cookie";
+
+export const initCSRFToken = async () => {
+  console.log("init CSRF token");
+  const targetUrl = targetURL(USER_AUTH_URL.concat(URLS_USER_AUTH.index));
+  await fetch(targetUrl, { mode: "cors" })
+    .then((response) => response.json())
+    .then((data) => {
+      if (data.result) {
+        localStorage.setItem("aiplay_csrf_token", data.csrf_token);
+        setCookie("csrftoken", data.csrf_token);
+      }
+    });
+};
 
 export const refreshToken = async () => {
   console.log("refresh token");
